Release 2.4.0-BETA
See #8643

Hi guys

I've created this beta release with some changes that 'should' ensure the childtickets data is now correctly cached and synced using the trac 0.12 'cached' decorator. I don't have much of a test env here at home to see if this 'really' works but as far as I can tell it's regenerating the cache when required (ticket add, mod, delete) and the decorator should ensure it is being done so across all threads.

I'd appreciate it if you could try it out and see if it also works with the Mylyn plugins.

Many Thx
Mark.



git-svn-id: http://trac-hacks.org/svn/childticketsplugin@10110 7322e99d-02ea-0310-aa39-e9a107903beb